Extract snippets to module
