+++ src/pages/AuthKitDemo.tsx
@@ -11,6 +11,7 @@ import { CodeBlock, atomOneDark } from 'react-code-blocks'
 import ConnectedWalletLabel from 'src/components/connected-wallet-label/ConnectedWalletLabel'
 import SafeInfo from 'src/components/safe-info/SafeInfo'
 import { useAccountAbstraction } from 'src/store/accountAbstractionContext'
+import { WEB3AUTH_SNIPPET } from 'src/utils/snippets'
 
 const AuthKitDemo = () => {
   const { loginWeb3Auth, isAuthenticated, safeSelected, chainId } = useAccountAbstraction()
@@ -100,7 +101,7 @@ const AuthKitDemo = () => {
       {/* TODO: create a component for this? */}
       <CodeContainer>
         <CodeBlock
-          text={code}
+          text={WEB3AUTH_SNIPPET}
           language={'javascript'}
           showLineNumbers
           startingLineNumber={96}
@@ -112,67 +113,6 @@ const AuthKitDemo = () => {
 }
 
 export default AuthKitDemo
-
-const code = `import { Web3AuthModalPack } from '@safe-global/auth-kit'
-import { OpenloginAdapter } from '@web3auth/openlogin-adapter'
-
-const options: Web3AuthOptions = {
-  clientId: process.env.REACT_APP_WEB3AUTH_CLIENT_ID,
-  web3AuthNetwork: 'testnet',
-  chainConfig: {
-    chainNamespace: CHAIN_NAMESPACES.EIP155,
-    chainId: chainId,
-    rpcTarget: rpcTarget
-  },
-  uiConfig: {
-    theme: 'dark',
-    loginMethodsOrder: ['google', 'facebook']
-  }
-}
-
-const modalConfig = {
-  [WALLET_ADAPTERS.TORUS_EVM]: {
-    label: 'torus',
-    showOnModal: false
-  },
-  [WALLET_ADAPTERS.METAMASK]: {
-    label: 'metamask',
-    showOnDesktop: true,
-    showOnMobile: false
-  }
-}
-
-const openloginAdapter = new OpenloginAdapter({
-  loginSettings: {
-    mfaLevel: 'mandatory'
-  },
-  adapterSettings: {
-    uxMode: 'popup',
-    whiteLabel: {
-      name: 'Safe'
-    }
-  }
-})
-
-const web3AuthModalPack = new Web3AuthModalPack({
-  txServiceUrl: 'https://safe-transaction-{chain}.safe.global',
-})
-
-await web3AuthModalPack.init({
-  options,
-  adapters: [openloginAdapter],
-  modalConfig
-})
-
-// Allow to login and get the derived EOA
-await web3AuthModalPack.signIn()
-
-// Logout
-await web3AuthModalPack.signOut()
-
-// Get the provider
-web3AuthModalPack.getProvider()
-`
 
 const ConnectContainer = styled(Box)<{
   theme?: Theme
